update 8 - auth works properly

+++ spring-mongo/src/main/java/com/linkshortener/controller/LinkController.java
@@ -3,8 +3,11 @@ package com.linkshortener.controller;
 import com.linkshortener.AppConstants;
 import com.linkshortener.entity.Link;
 
+import com.linkshortener.entity.User;
 import com.linkshortener.repository.LinkRepository;
 
+import com.linkshortener.repository.UserRepository;
+import com.linkshortener.security.services.UserDetailsImpl;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -15,16 +18,16 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.*;
 
 
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.nio.charset.Charset;
+import java.util.*;
 
 @RestController
 @RequestMapping("/api/link")
@@ -32,6 +35,9 @@ public class LinkController {
 
     @Autowired
     private LinkRepository linkRepository;
+
+    @Autowired
+    UserRepository userRepository;
 
     @Autowired
     private MongoTemplate mongoTemplate;
@@ -46,9 +52,14 @@ public class LinkController {
 
     @PreAuthorize("hasRole('USER')")
     @GetMapping("/getMyLinks")
-    public ResponseEntity<List<Link>> getMyLinks(@RequestBody String userID) {
+    public ResponseEntity<List<Link>> getMyLinks() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+
+
         Query query = new Query();
-        query.addCriteria(Criteria.where("belongs_to").is(userID));
+        query.addCriteria(Criteria.where("belongs_to").is(userDetails.getId()));
         List<Link> links = mongoTemplate.find(query, Link.class);
 
         return ResponseEntity.ok(links);
@@ -56,8 +67,25 @@ public class LinkController {
     }
 
     @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
-    @PostMapping("/add")
-    public ResponseEntity<Link> addLink(@RequestBody Link link) {
+    @PostMapping("/add/{url}")
+    public ResponseEntity<Link> addLink(@RequestParam String url) {
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+
+        byte[] array = new byte[6]; // length is bounded by 7
+        new Random().nextBytes(array);
+        String refString = new String(array, Charset.forName("UTF-8"));
+
+        Link link = new Link(
+                url,
+                refString,
+                userDetails.getId(),
+                0
+        );
+
+
+
         Link result = linkRepository.save(link);
         return ResponseEntity.ok(result);
     }
@@ -79,11 +107,17 @@ public class LinkController {
 
     @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
     @DeleteMapping("/deleteMyLink/{id}")
-    public ResponseEntity<Map<String, Object>> deleteMyLink(@RequestBody String userID, @PathVariable String id) {
+    public ResponseEntity<Map<String, Object>> deleteMyLink(@PathVariable String id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+
+        Optional<User> user = userRepository.findById(userDetails.getId());
+
         Optional<Link> link = linkRepository.findById(id);
 
 
-        if(link.isPresent() && link.get().getId().equals(id)){
+        if(link.isPresent() && link.get().getId().equals(id) && link.get().getBelongs_to().equals(user.get().getId())){
             linkRepository.deleteById(id);
 
             Map<String, Object> userMap = new HashMap<>();
@@ -104,6 +138,16 @@ public class LinkController {
     }
     @RequestMapping("/redirect/{ref}")
     public ResponseEntity<Link> redirect(@PathVariable String ref) throws URISyntaxException {
+
+        Query query = new Query();
+        query.addCriteria(Criteria.where("ref").is(ref));
+        Link link = mongoTemplate.findOne(query, Link.class);
+        if(link != null){
+            link.setClick(link.getClick() + 1);
+            linkRepository.save(link);
+        }
+
+
         URI yahoo = new URI(AppConstants.baseUrl+"/redirect/"+ref);
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.setLocation(yahoo);
